Adds support to fetch nested properties from objects and arrays.
To fetch a nested property, simply use a `.` separator:

    monocle.get('/users/123', {
      props: ['displayName', 'primaryPhoto.urlTemplate']
    })
    .then(function(user) {
      user.displayName; // 'Bob';
      user.primaryPhoto.urlTemplate; // 'http://...'
      user.primaryPhoto.photoId; // undefined
    });

To specify which properties to pluck from an array of objects, use the `@` prefix:

    monocle.get('/users', {
      props: ['@displayName', '@primaryPhoto.urlTemplate']
    })
    .then(function(users) {
      users[0].displayName; // 'Bob';
      users[0].primaryPhoto.urlTemplate; // 'http://...'
      users[0].primaryPhoto.photoId; // undefined
    });

**Why the `@` prefix for arrays?**

This helps make it clear that you're expecting an array of items to be returned, and that the following property should be plucked. Also, the character is easy to type and can be used in a browser without too much trouble (unlike `#`).