WTF-15958 Allow symlinks to be prehydrated to optimize resolution.
When creating a `Symlink`, you can now pass in prehydrated data as a POJO or a `Resource` instance. When the symlink is being resolved, it will first inspect the prehydrated data to optimize the props it requests from the underlying call.

```js
router.route('/users', usersSchema, {
  get: function(request, connection) {
    return [
      new Symlink('/users/1', { displayName: 'Bob' },
      new Symlink('/users/2', { displayName: 'Alice' },
      /* ... */
    ];
  }
});
```

When the symlink is being resolved, all requested properties that are found in the prehydrated data will be removed from the underlying request. If all requested properties are prehydrated, the underlying API call will not be made at all.